fix(auth,agent): resp.getCode() 跟 ResultCode 比较加 .intValue() 显式拆箱
改 2 处:
  ai-platform-auth/src/main/java/.../AuthService.java:105
  ai-platform-agent/src/main/java/.../ReActEngine.java:117

原因:
  resp.getCode() 返回 Integer, ResultCode.SUCCESS.getCode() 返回 Integer
  两者都是 Integer, != 比较会拆箱成 int, 如果 resp.getCode() 是 null
  (前面 == null 短路已经防住) 不会 NPE, 但显式 .intValue() 更清晰

  之前一直用 != Integer 的写法, 现在统一加 .intValue() 避免任何拆箱歧义

格式:  resp.getCode() != ResultCode.SUCCESS.getCode().intValue()

+++ backend/ai-platform-agent/src/main/java/com/aiplatform/agent/engine/ReActEngine.java
@@ -114,7 +114,7 @@ public class ReActEngine {
                 "maxTokens", 512
         );
         Result<Map<String, Object>> resp = inferenceClient.generate(body);
-        if (resp == null || resp.getCode() == null || resp.getCode() != ResultCode.SUCCESS.getCode()) {
+        if (resp == null || resp.getCode() == null || resp.getCode() != ResultCode.SUCCESS.getCode().intValue()) {
             // Fallback to a deterministic stub so the loop can still be exercised end-to-end
             // when the inference service is offline.
             return "{\"action\":\"final\",\"answer\":\"(本地模型未就绪) 已收到你的输入: " + escape(prompt) + "\"}";
